requirements.txt
pyautogui==0.9.54
opencv-python==4.8.1.78
Pillow==10.0.1
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pytesseract==0.3.10
openai==1.3.7
python-dotenv==1.0.0
requests==2.31.0
pydrive==1.3.1 